3D-77: Split shared profiling facade
