+++ src/shared/profiling/artifacts.py
@@ -1,0 +1,15 @@
+"""Artifact writers for profiling and diagnostic scripts."""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+from typing import Any
+
+
+def write_json(path: str | Path, payload: Any) -> Path:
+    """Write indented JSON and return the resolved ``Path`` object."""
+    out_path = Path(path)
+    out_path.parent.mkdir(parents=True, exist_ok=True)
+    out_path.write_text(json.dumps(payload, indent=2), encoding="utf-8")
+    return out_path
